dispatcher/requirements.txt
gitpython==3.1.43
python-frontmatter==1.1.0
pyyaml==6.0.2
pydantic==2.9.2
pydantic-settings==2.7.1
docker==7.1.0
slack-bolt==1.21.3
